Split MIR expression lowering
